the fast one
For some weird reason this version of the code seems to be very fast, the snake just zips off the screen.

I think it is somthing to do with how python handels loding variables into ram since it happpend when I was globalizing some.

minor fixes

added a feature where you have a one to 200 chance of getting a strabberry instead of a regualr block.

added a delay that needs to be fine tuned.

next version here I come.